demo/requirements.txt
numpy>=1.23
Pillow>=9.5
matplotlib>=3.8
